requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
PyYAML>=6.0
pytz>=2024.1
python-dotenv>=1.0.0
